backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
sqlalchemy==2.0.36
asyncpg==0.29.0
alembic==1.13.3

# Settings & validation
pydantic[email]==2.9.2
pydantic-settings==2.5.2

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

# Async HTTP (gigachat.py, yandexgpt.py)
httpx==0.27.2

# Task queue
celery==5.4.0
redis==5.1.1

# Scraping Алисы
playwright==1.47.0
tenacity==9.0.0

# Logging
structlog==24.4.0

# AI APIs
anthropic==0.34.2

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
